feat: add phone number conversions
